FIX #14 Multiple random messages
