+++ lib/defaultConfig.js
@@ -1,0 +1,9 @@
+const defaultConfig = {
+  requestInfoReplyComment: 'The maintainers of this repository would appreciate it if you could provide more information.',
+  requestInfoOn: {
+    issue: true,
+    pullRequest: true
+  }
+}
+
+module.exports = defaultConfig
